Move custom_where logic from numpy_wrapper to __init__
