+++ autograd/numpy/numpy_wrapper.py
@@ -190,32 +190,3 @@ else:
     @primitive
     def _astype(A, dtype, order="K", casting="unsafe", subok=True, copy=True):
         return A.astype(dtype, order, casting, subok, copy)
-
-
-# ----- Overwrite where to preserve native Python types -----
-_original_where = globals().get("where")
-
-
-def _is_np_or_autograd(x):
-    from autograd.tracer import Box
-
-    if isinstance(x, Box) or isinstance(x, (_np.ndarray, _np.generic)):
-        return True
-    return False
-
-
-def custom_where(condition, *args):
-    if not args:
-        return _original_where(condition)
-    if len(args) != 2:
-        raise ValueError("either both or neither of x and y should be given")
-    x, y = args
-    if _is_np_or_autograd(condition) or _is_np_or_autograd(x) or _is_np_or_autograd(y):
-        return _original_where(condition, x, y)
-    elif condition:
-        return x
-    else:
-        return y
-
-
-where = custom_where
